+++ spring-cloud-tencent-examples/quickstart-example/quickstart-caller-service/src/main/java/com/tencent/cloud/quickstart/caller/CustomMetadataProvider.java
@@ -28,8 +28,8 @@ import org.springframework.stereotype.Component;
 @Component
 public class CustomMetadataProvider implements InstanceMetadataProvider {
 
-	@Override
-	public String getRegion() {
-		return "huadong";
-	}
+//	@Override
+//	public String getRegion() {
+//		return "huadong";
+//	}
 }