+++ spring-cloud-tencent-polaris-context/src/main/java/com/tencent/cloud/polaris/context/tsf/metadata/TsfInstanceMetadataProvider.java
@@ -74,4 +74,14 @@ public class TsfInstanceMetadataProvider implements InstanceMetadataProvider {
 			}
 		}};
 	}
+
+	@Override
+	public String getRegion() {
+		return tsfCoreProperties.getTsfRegion();
+	}
+
+	@Override
+	public String getZone() {
+		return tsfCoreProperties.getTsfZone();
+	}
 }